drivers: watchdog: Allow saving/reading a retained value
Some watchdog controllers have the ability to store a value that is
retained between resets, if they were caused by the watchdog timer.

This can be used by various different purposes, including, but not
limiting to, tests.

This implements the feature for ESP32 and stubs all others.  A
return value of `0` signals success, and `-ENOSYS` is returned by
the stubs.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>