WIP: kernel: Implement k_futex()
This patch is a work-in-progress implementation of a futex for Zephyr.
The idea for futexes have been around for a while and was first
described in "Fuss, Futexes and Furworks: Fast Userlevel Locking in
Linux", by Franke and Russell[1].

The idea behind futexes is that it can, in most cases, remove the
need for a system call altogether: for instance, when implementing
a mutex, k_futex() needs to be called only if a CAS fails.

Not all operations from the Linux implementation of futex has been
ported over to Zephyr: only FUTEX_WAIT, FUTEX_WAKE.  No implementations
have been studied when writing this version; only man pages have been
consulted.

Futexes are looked up with a linear search; this isn't optimal and
will need to be revisited later.  However, for the low number of
channels for the moment (10), this is unlikely to be a big issue.

This code has not been tested, not even compiled.  I don't really
know if this actually useful in Zephyr.

[1] https://www.kernel.org/doc/ols/2002/ols2002-pages-479-495.pdf
[2] https://www.akkadia.org/drepper/futex.pdf